Split classes and add more checks
